Add sample launchd plist file
This file will work as-is provided clipper is in the standard PATH (eg.
/usr/bin/, but not /usr/local/bin or ~/bin). Alas, there is no way to
use ~, $HOME, or . (relative to $HOME) in the program name, which means
the tool needs to be installed globally.

To keep clipper running in the background at all times:

  cp com.wincent.clipper.plist ~/Library/LaunchAgents/
  launchctl load -w -S Aqua ~/Library/LaunchAgents/com.wincent.clipper.plist

One important thing to note: the above won't work inside a tmux session
because the process will end up running in the wrong execution context
and so won't have access to the clipboard.

To stop clipper:

  launchctl unload ~/Library/LaunchAgents/com.wincent.clipper.plist

If you want logging to a file, you can add StandardOutPath and
StandardErrorPath, but the same inability to specify paths relative to
$HOME is an announce here, so I'll likely add an option to clipper to
write to a specific logfile.

Signed-off-by: Wincent Colaiuta <win@wincent.com>