Add Homebrew-based install/uninstall info
Thanks to Shane da Silva (<http://shane.io/>), it's not possible to
install Clipper using Homebrew. As this is a fair bit easier than
installing manually (although it takes a long time if you haven't
previously installed apple-gcc42 and go packages, which end up
downloading about 185M of stuff) let's put it front and center in the
instructions.

Signed-off-by: Wincent Colaiuta <win@wincent.com>