Add section about Mosh to the docs
This commit adds some tips on getting Mosh, SSH and Clipper playing
nicely together. I include the section about the various `Control*`
settings because they are a complicating variable which needs to be
dealt with.

(In general, the `Control*` settings are a great time saver, but they
can make trouble-shooting difficult and we can simplify the business of
setting up a "pure-forwarder" for use with Mosh by keeping them out of
the picture.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>