Optionally listen on a UNIX domain socket
If `--address` (or "address" in the config file) looks like a path (ie. starts
with "/" or "~") then Clipper will listen on a UNIX domain socket at that path
rather than on a port.

Motivation: better security, because the socket can be protected via filesystem
permissions, whereas the TCP port is accessible to any user that has access to
localhost.

Netcat can be configured to talk to a UNIX socket (via `-U`), and it's possible
to forward local sockets using SSH as well, given a sufficiently recent version
of SSH, so everything can continue to work as before.

I'm going to run with this instead of my half-made passphrase-based solution,
for now, because the latter has a bunch of issues (need to exchange a shared
secret at start-time, and updating it may be tricky in the presence of
long-lived tmux sessions etc).

This is still a WIP, however, as we have some issues to work through. For
example, command-line options should trump config file options (we currently do
the opposite) and it all needs to be documented still. Also, the code is
frickin' ugly (it's been so long since I programmed in Go, I had to Google "how
do I X?" about 20 times while preparing this tiny change).