Split plist files into two
Homebrew has it's own custom fork:

  https://github.com/Homebrew/homebrew-core/blob/409a67fbf587e4346235cb7ee71eab0d65adc88f/Formula/clipper.rb

So I am not worried about breaking that, but I do want to show a couple of clear
examples of how to set this up via TCP ports or UNIX domain sockets.

Note that I also updated the paths to use `/usr/local/bin` instead of `/usr/bin`
or a relative PATH because of Apple's SIP (System Integrity Protection) that
makes it much less likely that people will be able to write to the latter.