Add --executable and --args flags
Alas, `--args` must break with the pattern of having the short option
start with the same letter as the long option (ie. `-a`); instead it
uses `-r`. This makes the `--help`/`-h` output a bit "meh"; not sure if
there's a way to fix that. If I can't find one I might need to rename
this option to `--flags`/`-f` or `--options`/`-o`.