Rename --args/-r to --flags/-f
I don't want to reimplement `flag.PrintDefaults` in order to get
appropriately grouped usage information, so I am renaming this flag to
so that the long and short forms both begin with the same letter.

"Flag" is unfortunately pretty overloaded at this point, and we already
have a bit of a cluster going on with "Settings", "Options", "Defaults",
and "Flags", but this is the way things are for now: because of the way
JSON marshalling works, we have to have names which match the desired
options.