Don't listen on all interfaces
This is the dumb fix for the face-palm bug reported here:

  https://github.com/wincent/clipper/issues/9

wherein our attempt to listening on IPv4 and IPv6 loopback interfaces
when available was actually causing us to listen on all interfaces
(ugh).

I'll remove the duplication in the next commit, and also try to
gracefully handle systems which might not have IPv6 interfaces up.