Bail if we failed to set up any listeners
Normally we only try to set up one listener.

But in the case where we try to listen on the IPv4 and IPv6 loopback
addresses, we're fine with up to one of those failing, but not both.
This `listenerCount` variable is ugly, but I don't know Go idioms. I
think I might just pull it out into a separate loop for clarity (in a
language with a generic `filter` function I would just use that to strip
out `nil` and then look at the `len`).