Add a filter function to keep main body cleaner
So I've now done this three ways to see which one sucks less. The lack
of generics makes this one feel a bit ridiculous, but at least the main
body reads nicely now.

Note also the pass-by-reference shennanigans to avoid making a copy of
the listener objects (I don't even know what that would mean, although I
want to try it out and find out: maybe Go only makes a copy of the
slice, not the slice items, when you pass a slice by value).