doc: make it clear that StreamLocalBindUnlink should be set on the server
I've never found it to work on the client, but it does work on the
server. It wasn't clear to me from the docs that this would be the case
because the wording is almost identical for both `sshd_config`:

  StreamLocalBindUnlink
          Specifies whether to remove an existing Unix-domain
          socket file for local or remote port for- warding before
          creating a new one.  If the socket file already exists and
          StreamLocalBindUnlink is not enabled, sshd will be unable to
          forward the port to the Unix- domain socket file.  This option
          is only used for port forwarding to a Unix-domain socket file.

          The argument must be yes or no.  The default is no.

and `ssh_config`:

  StreamLocalBindUnlink
          Specifies whether to remove an existing Unix-domain
          socket file for local or remote port for- warding before
          creating a new one.  If the socket file already exists and
          StreamLocalBindUnlink is not enabled, ssh will be unable to
          forward the port to the Unix- domain socket file.  This option
          is only used for port forwarding to a Unix-domain socket file.

          The argument must be yes or no (the default).

I'm thinking that perhaps the client side config only applies to
port-forwarding (or the socket on the local side), and the server config
applies to reverse port forwarding (or the the socket on the remote
side).