Overhaul flag/config handling to deal with empty values
We were previously conflating non-emptiness of flags and config values
with being-set. This meant that you couldn't explicitly set a value to
an empty string, because we would interpret that as the value not being
set and we would fall back to the default. This is a problem if you
want, for example, to suppress all `--flags` being sent to the
underlying `--executable`.

Fix this by replacing our simple string and int variables with struct
wrappers that contain both the value and an explicit boolean that
indicates whether the value was provided. This requires us to implement
the `flag.Value` interface on our wrapper structs, and also the
`json.Unmarsheler` interface so that we can unpack from the JSON config
file into these custom structs.

In addition to fixing the issue, these changes make the code in the
`mergeSettings()` function look simpler and more fool-proof now. Only
downside is we need a bit of boilerplate-ish code to implement those
interfaces. Error messages when providing malformed JSON are about as
good as they used to be:

    clipper: 2018/10/04 13:15:31 json: cannot unmarshal object into Go
    struct field Options.Address of type string

Closes: https://github.com/wincent/clipper/issues/15