refactor: make log file directory on all platforms
I don't see any reason why this should be limited to Linux, so remove
the platform check.

Additionally, check the error condition and blow up if we can't make the
directory. Finally, remove "config" from the variable name because it
may not be a `~/.config` directory at all.

And finally, avoid calling `expandPath` twice.