feat: add `brew` target to Makefile
So that I don't have to look up how to do this next time around.

Creates a PR like:

- https://github.com/Homebrew/homebrew-core/pull/152828

Although note, it may appear to hang of a long time; at least for me, it
wanted to upload a zillion objects to my homebrew-core fork before
creating the PR, and that took about 12 minutes.